lecture-7-hw-2
• Reorganized file structure to have a components folder
• Improved RTK data usage. (from props drilling to selectors)
