+++ src/components/counter/component.jsx
