+++ src/components/menu/component.jsx
@@ -2,15 +2,16 @@ import MenuItem from "../menuItem/component";
 
 import styles from './style.module.css'
 
-function Menu({menu}) {
+
+function Menu({menuDishesIds}) {
 
     return (  
     <div className={styles.menuWrapper}>
         <h3>Меню</h3>
-        {menu?.length && (
+        {menuDishesIds?.length && (
             <ul className={styles.listWrapper}>
-            {menu.map((element) => {
-                return <MenuItem menuItem={element} />
+            {menuDishesIds.map((dishId) => {
+                return <MenuItem dishId={dishId} />
             })}
         </ul>
         )}
