+++ src/components/restarauntWrapper/component.jsx
@@ -2,7 +2,7 @@ import { useSelector } from "react-redux";
 import RestaurantTabsContainer from "../restarauntTabsContainer/component";
 import Restaurant from "../restauraunt/component";
 import { useState, useEffect } from "react";
-import { selectRestarantsIds, selectRestarauntById } from "../redux/entities/restaraunt";
+import { selectRestarantsIds, selectRestarauntById } from "../../redux/entities/restaraunt";
 
 function RestaurantWrapper() {
     const restaurants = useSelector(selectRestarantsIds);
@@ -10,8 +10,6 @@ function RestaurantWrapper() {
     // Set the initial state to the first restaurant's ID
     const [currentRestaurantId, setCurrentRestaurantId] = useState(restaurants[0]);
 
-    // Use the currentRestaurantId to select the current restaurant from the Redux store
-    const currentRestaurant = useSelector((state) => selectRestarauntById(state, currentRestaurantId));
 
     // Update the currentRestaurantId when restaurants array changes (e.g., after data is loaded)
     useEffect(() => {
@@ -22,8 +20,8 @@ function RestaurantWrapper() {
 
     return (
         <>
-            <RestaurantTabsContainer currentRestaurant={currentRestaurant} setCurrentRestaurantId={setCurrentRestaurantId} />
-            <Restaurant currentRestaurant={currentRestaurant} />
+            <RestaurantTabsContainer currentRestaurantId={currentRestaurantId} setCurrentRestaurantId={setCurrentRestaurantId} />
+            <Restaurant currentRestaurantId={currentRestaurantId} />
         </>
     );
 }
